#13 [mod] 선택된 테마 넘기기 로직 수정
