+++ app/src/main/java/com/sopetit/softie/ui/onboarding/themechoice/ChoiceThemeViewModel.kt
@@ -92,15 +92,7 @@ class ChoiceThemeViewModel : ViewModel() {
     val themeBtnEnabled: LiveData<Boolean>
         get() = _themeBtnEnabled
 
-    private val _selectedThemeArray: MutableLiveData<ArrayList<Int>> = MutableLiveData()
-    val selectedThemeArray: LiveData<ArrayList<Int>>
-        get() = _selectedThemeArray
-
     fun setThemeBtnEnabled(boolean: Boolean) {
         _themeBtnEnabled.value = boolean
-    }
-
-    fun setSelectedThemeArray(themeArray: ArrayList<Int>) {
-        _selectedThemeArray.value = themeArray
     }
 }
